# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YU-GI-OH! Api Rest",
        "description": "Implementation of Yu-Gi-Oh! Api Rest",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local enviroment"
        }
    ],
    "paths": {
...
        "/api/subtypes": {
            "get": {
                "tags": [
                    "Subtypes"
                ],
                "summary": "List all subtypes.",
                "operationId": "subtype-index",
                "responses": {
                    "200": {
                        "description": "Get a subtypes collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubtypeResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Subtypes"
                ],
                "summary": "Create a subtype.",
                "operationId": "subtype-store",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Subtype's name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subtype created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubtypeResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/subtypes/{id}": {
            "get": {
                "tags": [
                    "Subtypes"
                ],
                "summary": "Get a specific subtype.",
                "operationId": "subtype-show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subtype ID.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subtype obteined.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubtypeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Subtypes"
                ],
                "summary": "Update a subtype.",
                "operationId": "subtype-update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subtype ID.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Subtype's name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subtype updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubtypeResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SubtypeResource": {
                "title": "Subtype",
                "description": "Card's subtype",
                "properties": {
                    "id": {
                        "description": "Subtype's ID",
                        "type": "integer",
                        "format": ""
                    },
                    "name": {
                        "description": "Subtype's name",
                        "type": "string",
                        "format": ""
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Normal Monster"
                }
            }
        }
    }
}